json-to-index/requirements.txt
azure-functions>=1.12.0
azure-storage-blob>=12.0.0
azure-search-documents>=11.4.0
azure-core>=1.26.4